refactor services to be a bit more consistent

+++ backend/controllers/importController.js
@@ -8,29 +8,33 @@ export class ImportController {
 
     static async fileUpload(req, res) {
 
+        const file = req.file;
+        if (!file) {
+            return res.status(400).json({ message: "Please upload a file!" });
+        }
         try {
-            const file = req.file;
-            console.log(req.from);
             let output = [];
-            let readStream = await fs.createReadStream(file.path).pipe(csv.parse({headers: true})).on('data', (row) => {
+            let readStream = fs.createReadStream(file.path).pipe(csv.parse({ headers: true })).on('data', (row) => {
                 output.push(row);
             });
 
             readStream.on('end', async () => {
-                await fs.unlinkSync(file.path);
+                fs.unlinkSync(file.path);
                 try {
                     await Validator.validateFile(output);
                 } catch (e) {
-                    return res.status(400).json({message: e.message});
+                    return res.status(400).json({ message: e.message });
                 }
                 const token = req.headers.authorization;
                 const decoded = jwt.verify(token, process.env.JWT_SECRET);
-                ImportService.addFileData(output, decoded.userId);
-                return res.status(201).json({message: "File uploaded successfully"});
+                console.log(decoded.userId);
+
+                let response = await ImportService.addFileData(output, decoded.userId);
+                return res.status(response.code).json({ message: response.message });
             });
 
         } catch (error) {
-            return res.status(500).json({message: error.message});
+            return res.status(500).json({ message: error.message });
         }
     }
